On Windows running salt with admin privileges now authenticates
On Windows, salt.utils.get_specific_user will now return
'sudo_<username>' when the the client process has administrator
privileges (eg used "Run as administrator" option).

This enables communication with the master if the client and
salt-master processes are run as different users (which is a
common use case when salt-master is run as a service and the
'salt' command is run in an administrator cmd.exe under the
logged in user).

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>