Windows: salt fails to start when running as domain user (#33153)
The following error occurs when running salt-master or salt-minion
under a domain user account:

  File "...\lib\site-packages\
  salt\utils\win_functions.py", line 79, in get_user_groups
      groups = win32net.NetUserGetLocalGroups(None, name)
      pywintypes.error: (2221, 'NetUserGetLocalGroups',
      'The user name could not be found.')

`salt/utils/__init__.py`:
- Change `get_user` to use:
`win32api.GetUserNameEx(win32api.NameSamCompatible)`. This will return
the username in the format 'DOMAIN\username'.

`salt/utils/win_functions.py`:
- Change `get_current_user` to also use
`win32api.GetUserNameEx(win32api.NameSamCompatible)`. This will fix the
problem when it passes the username to `win32net.NetUserGetLocalGroups`.

`salt/client/__init__.py`:
- In `LocalClient.__read_master_key`, on Windows replace the username
portion of the keyfile path so that '\' is converted to '_'. This will
allow a valid filename.

`salt/daemons/masterapi.py`:
- In `access_keys`, on Windows replace the username
portion of the keyfile path so that '\' is converted to '_'. This will
allow a valid filename.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>